+++ CHANGELOG.md
@@ -131,3 +131,7 @@ This update adds the ability to interface with external services to enable new f
     - Aircraft with no location data are now filtered from the alert display.
     - Improved the resiliency of the data intake process.
     - Fixed an issue where the message intake file could grow exponentially.
+- Improved drone alert process.
+    - The wireless threat database can now contain full MAC addresses to identify individual devices.
+    - The drone detection process now considers all Airodump-NG output files, not just the first.
+    - Automatic start-up mode is not significantly more reliable.